refactor: move processing logic from EpisodeProvider to BasicEpisodeParser
